Fix JMXFetch issue when several instances on the same host (#124)
* Remove storage of connections in the connection manager
The connection will now be stored by the instance. No more shared connections

* Change ConnectionManager into ConnectionFactory
